% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT3documentation.R
\name{GAPIT3documentation}
\alias{GAPIT3documentation}
\title{GAPIT3documentation}
\usage{
GAPIT3documentation()
}
\value{
a list
}
\description{
GAPIT3documentation
}
\examples{
GAPIT3documentation()

}
